+++ bank/src/main/java/com/example/bank/model/ProductTypeModel.java
@@ -3,6 +3,7 @@ package com.example.bank.model;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonValue;
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
@@ -75,7 +76,7 @@ public class ProductTypeModel {
   private DescriptionEnum description = DescriptionEnum.SAVING;
 
   @JsonProperty("maintenance_fee")
-  private Integer maintenanceFee;
+  private BigDecimal maintenanceFee;
 
   @JsonProperty("monthly_transaction_limit")
   private Integer monthlyTransactionLimit;
@@ -84,7 +85,7 @@ public class ProductTypeModel {
   private Integer transactionDay;
 
   @JsonProperty("opening_amount")
-  private Integer openingAmount;
+  private BigDecimal openingAmount;
 
   @JsonProperty("is_delete")
   private Boolean isDelete = false;
