+++ bank/src/main/java/com/example/bank/model/dto/CustomerProductDto.java
@@ -1,6 +1,7 @@
 package com.example.bank.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
@@ -39,10 +40,10 @@ public class CustomerProductDto {
   private CardDto numberCard;
 
   @JsonProperty("amount")
-  private Integer amount;
+  private BigDecimal amount;
   
   @JsonProperty("credit_limit")
-  private Integer creditLimit;
+  private BigDecimal creditLimit;
 
   @JsonProperty("number_transaction_limit")
   private Integer numberTransactionLimit;
