+++ bank/src/main/java/com/example/bank/model/dto/OperationDto.java
@@ -1,6 +1,7 @@
 package com.example.bank.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
@@ -30,10 +31,10 @@ public class OperationDto {
   private CustomerProductDto customerProducto;
 
   @JsonProperty("balance")
-  private Integer balance;
+  private BigDecimal balance;
 
   @JsonProperty("commission")
-  private Integer commission;
+  private BigDecimal commission;
 
   @JsonProperty("operation_date")
   private String operationDate;
